initial push
If the type does NOT have use in phred and was generated by steven, it was removed.

If the type does NOT have a use in phred but was generated by an MSI (for a project that has not yet landed in phred) it was left the same.

Otherwise, the type was renamed and tried to conform to the base type + NONCANONICAL_#. Some inconsistencies since some devices are called NON_CANONICAL_# but those are for new sites that havent yet landed in phred, so this cleanup will be left until later.