fix for: Check multistate values are not empty bug: b/264439914
summary: It is desired to update the instance validator to catch when multistates are not corectly mapped. A real example given,

```
translation:
    level_status:
      present_value: points.dwst_1_tank_level_status.present_value
      states:
        MEDIUM: true
        LOW:
        HIGH: false
```

LOW is not mapped to any

As such it is desired not to instantiate such an entity to begin with as it isn't valid. To do so entity_instance.py _ParseTranslation function was modified to include additional logic that checks that the values for each multistate key/value pair have a non-null value that is neither None or '' (the empty string). it a null value is detected a ValueError is raised. A test has also been included

PiperOrigin-RevId: 502603669