fix for: linkage error not detected bug: b/264558814
summary: DB API (Java validator) catches a missing field in the source entity defined in a link whereas the Instance Validator (Python validator) does not. Investigating the bug we discovered that the _LinksAreValid method in entity_instance.py did not perform the required check on the source entity for the source field name; it only checked that link types are valid according to the entity type in the ontology. We added a check to validate that the source entity contains the source field referenced in the target field on a link. We also included a test taken from the building configuration which surfaced the bug.
PiperOrigin-RevId: 503251819