fix for: [Instance Validator] Check for states or units on a field bug: b/266246501
summary: We check to ensure that states or units are present on an entity via its Validate method via the _ValidateTranslation method. In there we included a check to see if the field translation is an instance of NonDimensionalValue; if so, print error (as non-dimensional values have a field name but not units or states) and setting is_valid to False and continuing with remaining validation.

Additionally did some refactoring and moved the parsing of entity operation and update_mask into _ParseOperationAndUpdateMask

We updated the entity_instance_test setUp() method to build to full ontology instead of the simplified universe. This was done to capture the SAFETY/FACP_FA2X real life test example (with some details modified for data security)

PiperOrigin-RevId: 504667084