fix for: [Instance Validator] Entities with Cloud device id must have translations bug: b/266437564
summary: it is desired to have entities with cloud_device_id(s) contain translations as well. there were two places we thought to implement: 1. fromYaml or 2. Validate method. for (1) the benefit is avoidance of creating an invalid instance (as our main interaction with instantiation of an entity is through its fromYaml method). However, we opted to implement in (2) as it is more generally falls inline with validation of an instance rather than its creation. If this isn't the case the implementation can be changed :)
PiperOrigin-RevId: 506412628